Flask/deployment/requirements_for_serverless_GCPbucket.txt
requests==2.28.1
numpy==1.21.6
scikit-learn==1.0.2
google-cloud-storage==1.17.0